+++ src/System.Private.CoreLib/src/System/Diagnostics/Tracing/EventPipeController.cs
@@ -1,0 +1,11 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Diagnostics.Tracing
+{
+    sealed internal class EventPipeController
+    {
+        public static bool Initializing { get => true; }
+    }
+}
